Introduction to Programming
This project was created in Java using the Eclipse IDE. 
The Inventory_Management_System_Project.zip file contains the Store Class and Store GUI classes in the src folder. In addition, there is the StoreGUICheck class in the inventory management folder, where error handling improvements were made to the StoreGUI class.